Refactored subsetting
