+++ data/netcdf_data.py
@@ -1,12 +1,17 @@
 from netCDF4 import Dataset, netcdftime
+from flask_babel import format_date
 from data.data import Data, Variable, VariableList
+from oceannavigator.nearest_grid_point import find_nearest_grid_point
 import xarray as xr
+import os
 from cachetools import TTLCache
 import pytz
 import warnings
 import pyresample
 import numpy as np
 import re
+import geopy
+import pandas
 
 class NetCDFData(Data):
 
@@ -24,6 +29,114 @@ class NetCDFData(Data):
 
     def __exit__(self, exc_type, exc_value, traceback):
         self._dataset.close()
+
+    """
+
+    """
+    def subset(self, query):
+        
+        # Ensure we have an output folder that will be cleaned by tmpreaper
+        if not os.path.isdir("/tmp/subset"):
+            os.makedirs("/tmp/subset")
+        working_dir = "/tmp/subset/"
+        
+        entire_globe = True # subset the globe?
+        if 'min_range' in query:
+            # Area explicitly specified
+            entire_globe = False
+            # Bounding box extents
+            bottom_left = [float(x) for x in query.get('min_range').split(',')]
+            top_right = [float(x) for x in query.get('max_range').split(',')]
+
+        # Time range
+        time_range = [int(x) for x in query.get('time').split(',')]
+        apply_time_range = False
+        if time_range[0] != time_range[1]:
+            apply_time_range = True
+
+        # Finds a variable in a dictionary given a substring containing common characters
+        # Not a fool-proof method but I want to avoid regex because I hate it.
+        variable_list = list(self._dataset.variables.keys())
+        def find_variable(substring):
+            for key in variable_list:
+                if substring in key:
+                    return key
+        
+        def get_coords():
+            
+            print("test")
+
+        # Get lat/lon variable names from dataset (since they all differ >.>)
+        lat = find_variable("lat")
+        lon = find_variable("lon")
+
+        if not entire_globe:
+            # Find closest indices in dataset corresponding to each calculated point
+            ymin_index, xmin_index, _ = find_nearest_grid_point(
+                bottom_left[0], bottom_left[1], self._dataset, self._dataset.variables[lat], self._dataset.variables[lon]
+            )
+            ymax_index, xmax_index, _ = find_nearest_grid_point(
+                top_right[0], top_right[1], self._dataset, self._dataset.variables[lat], self._dataset.variables[lon]
+            )
+
+            y_slice = slice(ymin_index, ymax_index)
+            x_slice = slice(xmin_index, xmax_index)
+
+            # Get nicely formatted bearings
+            p0 = geopy.Point(bottom_left)
+            p1 = geopy.Point(top_right)
+        else:
+            y_slice = slice(self._dataset.variables[lat].size)
+            x_slice = slice(self._dataset.variables[lon].size)
+
+            p0 = geopy.Point([-85.0, -180.0])
+            p1 = geopy.Point([85.0, 180.0])
+
+        # Get timestamp
+        time_variable = find_variable("time")
+        timestamp = str(format_date(pandas.to_datetime(np.float64(self._dataset[time_variable][time_range[0]].values)), "yyyyMMdd"))
+        endtimestamp = ""
+        if apply_time_range:
+            endtimestamp = "-" + str(format_date(pandas.to_datetime(np.float64(self._dataset[time_variable][time_range[1]].values)), "yyyyMMdd"))
+        
+        dataset_name = query.get('dataset_name')
+        # Do subsetting
+        if "riops" in dataset_name:
+            # Riops has different coordinate names...why? ¯\_(ツ)_/¯
+            subset = self._dataset.isel(yc=y_slice, xc=x_slice)
+        elif dataset_name == "giops_forecast":
+            subset = self._dataset.isel(latitude=y_slice, longitude=x_slice)
+        else:
+            subset = self._dataset.isel(y=y_slice, x=x_slice)
+
+        # Select requested time (time range if applicable)
+        if apply_time_range:
+            subset = subset.isel(**{time_variable: slice(time_range[0], time_range[1] + 1)}) # slice doesn't include the last element
+        else:
+            subset = subset.isel(**{time_variable: slice(int(time_range[0]), int(time_range[0]) + 1)})
+
+        # Filter out unwanted variables
+        output_vars = query.get('variables').split(',')
+        output_vars.extend([find_variable("depth"), time_variable, lat, lon]) # Keep the coordinate variables
+        for variable in subset.data_vars:
+            if variable not in output_vars:
+                subset = subset.drop(variable)
+
+        output_format = query.get('output_format')
+        filename =  dataset_name + "_" + "%dN%dW-%dN%dW" % (p0.latitude, p0.longitude, p1.latitude, p1.longitude) \
+                    + "_" + timestamp + endtimestamp + "_" + output_format
+
+        # Save subset normally
+        subset.to_netcdf(working_dir + filename + ".nc", format=output_format)
+
+        if int(query.get('should_zip')) == 1:
+            myzip = zipfile.ZipFile('%s%s.zip' % (working_dir, filename), mode='w')
+            myzip.write('%s%s.nc' % (working_dir, filename), os.path.basename('%s%s.nc' % (working_dir, filename)))
+            myzip.comment = 'Generated from www.navigator.oceansdata.ca'
+            myzip.close() # Must be called to actually create zip
+            return working_dir, filename+".zip"
+
+        return working_dir, filename+".nc"
     
     """
         Interpolates data given input and output definitions
